requirements.txt
mysql-connector-python==8.4.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.1
PyQt6-sip==13.6.0
